fix: Split code for endpoints by resource to improve maintainability
