+++ src/order_book_simulator/gateway/routers/orders.py
@@ -1,0 +1,38 @@
+from datetime import datetime, timezone
+from decimal import Decimal
+from uuid import uuid4
+
+from fastapi import APIRouter, Depends, HTTPException
+
+from order_book_simulator.common.models import OrderRequest, OrderResponse, OrderStatus
+from order_book_simulator.database.connection import get_db
+from order_book_simulator.gateway.app_state import app_state
+from order_book_simulator.gateway.validation import validate_order
+
+router = APIRouter(prefix="/orders", tags=["orders"])
+
+
+@router.post("", response_model=OrderResponse)
+async def create_order(order_request: OrderRequest, db=Depends(get_db)):
+    start_time = datetime.now(timezone.utc)
+    if app_state.producer is None:
+        raise HTTPException(
+            status_code=503, detail="Order processing service is unavailable"
+        )
+
+    async with db.begin():
+        await validate_order(order_request, db)
+
+        order_record = {
+            "id": uuid4(),
+            **order_request.model_dump(),
+            "status": OrderStatus.PENDING,
+            "filled_quantity": Decimal("0"),
+            "total_fee": Decimal("0"),
+            "gateway_received_at": start_time.isoformat(),
+        }
+
+        await app_state.producer.send_order(order_record)
+        return OrderResponse(
+            **order_record, created_at=start_time, updated_at=start_time
+        )
